Struct puffin::agent::Agent

pub struct Agent<PB: ProtocolBehavior> {
    descriptor: AgentDescriptor,
    put: Box<dyn Put<PB>>,
}

An Agent holds a non-cloneable reference to a Stream.

Fields
descriptor: AgentDescriptorput: Box<dyn Put<PB>>
Implementations

impl<PB: ProtocolBehavior> Agent<PB>

pub fn new(descriptor: AgentDescriptor, put: Box<dyn Put<PB>>) -> Self

pub fn progress(&mut self) -> Result<(), Error>

pub fn reset(&mut self, new_name: AgentName) -> Result<(), Error>

pub fn shutdown(&mut self) -> String

Shut down the agent by consuming it and returning a string that summarizes the execution.

pub fn is_state_successful(&self) -> bool

Checks whether the agent is in a good state.

pub fn is_reusable_with(&self, other: &AgentDescriptor) -> bool

Checks whether the agent is reusable with the descriptor.

pub fn name(&self) -> AgentName

pub fn put(&self) -> &dyn Put<PB>

pub fn put_mut(&mut self) -> &mut dyn Put<PB>