Module puffin::algebra::term

This module providesDYTermsas well as iterators over them.

Structs

Term

Terms are Terms equipped with optional Payloads when they no longer are treated as symbolic terms.

Enums

DYTerm

A first-order term: either a Variable or an application of an Function.

Constants

BITSTRING_NAME 🔒

SIZE_LEAF 🔒

Traits

Subterms

TermType

Trait for data we can treat as terms (either DYTerm or Term)

Functions

append 🔒

append_eval 🔒

display_term_at_depth 🔒

has_payload_to_replace_rec

remove_fn_prefix 🔒

remove_prefix 🔒

tlspuffin::term::op_impl::op_protocol_version -> op_protocol_version alloc::Vec<rustls::msgs::handshake::ServerExtension> -> Vec<rustls::msgs::handshake::ServerExtension>

Type Aliases

ConcreteMessage