Struct puffin::algebra::term::Term

pub struct Term<PT: ProtocolTypes> {
    pub term: DYTerm<PT>,
    pub payloads: Option<Payloads>,
}

Terms are Terms equipped with optional Payloads when they no longer are treated as symbolic terms.

Fields
term: DYTerm<PT>payloads: Option<Payloads>
Implementations

impl<PT: ProtocolTypes> Term<PT>

pub(crate) fn eval_until_opaque<PB>( &self, eval_tree: &mut EvalTree, ctx: &TraceContext<PB>, with_payloads: bool, sibling_has_payloads: bool, type_term: &TypeShape<PT>, ) -> Result<(Box<dyn EvaluatedTerm<PT>>, Vec<PayloadContext<'_, PT>>), Error>
where PB: ProtocolBehavior<ProtocolTypes = PT>,

Evaluate a term without replacing the payloads (returning the payloads with the corresponding paths instead, i.e., a Vec<PayloadContext> accumulator), except when reaching an opaque term with payloads as strict sub-terms. In the latter case, fully evaluate each of the arguments (and performing the payload replacements) before evaluating the opaque function, which then needs to be read to convert it back to a Box<dyn EvaluatedTerm<PT>>. @path: current path of &self in the overall recipe (initially []). Invariant: Returns the payloads to replace in this order: deeper first, left-most arguments first. When with_payloads is false, then this should be equivalent to evaluate_lazy_test and it always return empty PayloadContext vectors.

impl<PT: ProtocolTypes> Term<PT>

pub fn height(&self) -> usize

Height of term, considering non-symbolic terms as atoms

pub fn is_list(&self) -> bool

When the term starts with a list function symbol

pub fn is_opaque(&self) -> bool

When the term starts with an opaque function symbol (like encryption)

pub fn erase_payloads_subterms(&mut self, is_subterm: bool)

Erase all payloads in a term, including those under opaque function symbol

pub fn add_payload(&mut self, payload: Vec<u8>)

Add a payload at the root position, erase payloads in strict sub-terms not under opaque

pub fn make_payload<PB>(&mut self, ctx: &TraceContext<PB>) -> Result<(), Error>
where PB: ProtocolBehavior<ProtocolTypes = PT>,

Make and Add a payload at the root position, erase payloads in strict sub-terms not under opaque

pub fn all_payloads(&self) -> Vec<&Payloads>

Return all payloads contains in a term, even under opaque terms. Note that we keep the invariant that a non-symbolic term cannot have payloads in struct-subterms, see add_payload/make_payload.

pub fn all_payloads_mut(&mut self) -> Vec<&mut Payloads>

Return all payloads contains in a term (mutable references), even under opaque terms. Note that we keep the invariant that a non-symbolic term cannot have payloads in struct-subterms, see add_payload/make_payload.

pub fn payloads_to_replace(&self) -> Vec<&Payloads>

Return all payloads contained in a term, except those under opaque terms. The deeper the first in the returned vector.

pub fn has_payload_to_replace(&self) -> bool

Return whether there is at least one payload, except those under opaque terms.

pub fn has_payload_to_replace_wo_root(&self) -> bool

Return whether there is at least one payload, except those under opaque terms and at the root..

impl<PT: ProtocolTypes> Term<PT>

fn unique_id(&self, tree_mode: bool, cluster_id: usize) -> String

fn node_attributes( displayable: impl Display, color: &str, shape: &str, ) -> String

fn collect_statements( term: &Self, tree_mode: bool, cluster_id: usize, statements: &mut Vec<String>, )

pub fn dot_subgraph( &self, tree_mode: bool, cluster_id: usize, label: &str, ) -> String

If tree_mode is true then each subgraph is self-contained and does not reference other clusters or nodes outside of this subgraph. Therefore, only trees are generated. If it is false, then graphs are rendered.

impl<PT: ProtocolTypes> Term<PT>

pub fn count_functions_by_name(&self, find_name: &'static str) -> usize