Enum puffin::error::Error

pub enum Error {
    Fn(FnError),
    Term(String),
    Put(String),
    IO(String),
    Agent(String),
    Stream(String),
    Extraction(),
    SecurityClaim(&'static str),
}

Variants

Fn(FnError)

Returned if a concrete function from the module [tls] fails or term evaluation fails

Term(String)

Put(String)

PUT reported an error

IO(String)

There was an unexpected IO error. Should never happen because we are not fuzzing on a network which can fail.

Agent(String)

Some error which was caused because of agents or their names. Like an agent which was not found.

Stream(String)

Error while operating on a [Stream]

Extraction()

SecurityClaim(&'static str)