Module puffin::fuzzer::bit_mutations

Macros

expand_mutation 🔒

Structs

BitFlipMutatorDY

mutation definition

ByteAddMutatorDY

mutation definition

ByteDecMutatorDY

mutation definition

ByteFlipMutatorDY

mutation definition

ByteIncMutatorDY

mutation definition

ByteInterestingMutatorDY

mutation definition

ByteNegMutatorDY

mutation definition

ByteRandMutatorDY

mutation definition

BytesCopyMutatorDY

mutation definition

BytesDeleteMutatorDY

mutation definition

BytesExpandMutatorDY

mutation definition

BytesInsertCopyMutatorDY

BytesInsertMutatorDY

mutation definition

BytesLargeExpandMutator

Large number of bytes expand mutation for inputs with a bytes vector (expand range: from 2^5 to 2^12)

BytesLargeExpandMutatorDY

mutation definition

BytesRandInsertMutatorDY

mutation definition

BytesRandSetMutatorDY

mutation definition

BytesSetMutatorDY

mutation definition

BytesSwapMutatorDY

CrossoverInsertMutatorDY

CrossoverReplaceMutatorDY

DwordAddMutatorDY

mutation definition

DwordInterestingMutatorDY

mutation definition

MakeMessage

MAKE MESSAGE: transforms a sub term into a message which can then be mutated using havoc

QwordAddMutatorDY

mutation definition

ReadMessage

READ MESSAGE: picks a sub-term having itself a sub-term with payload, evaluate, read and performs an in-place replacement

SpliceMutatorDY

WordAddMutatorDY

mutation definition

WordInterestingMutatorDY

mutation definition

Functions

all_mutations

bit_mutations_dy

buffer_copy 🔒 ⚠

Mem move between vecs

buffer_self_copy 🔒 ⚠

Copied from libafl::mutators::mutations Mem move in the own vec

choose_payload_mut

choose_term_with_payload_mut

Returns the focused payload or randomly choose a non-symbolic payload in a trace (mutable reference).

havoc_mutations_dy

locate_diffs 🔒

Copied from libafl::mutators::mutations Returns the first and last diff position between the given vectors, stopping at the min len

make_message_term 🔒

MakeMessage on the term at path path in tr.

Type Aliases

AllMutations

BitMutations

HavocMutationsTypeDY