Trait puffin::protocol::EvaluatedTerm

pub trait EvaluatedTerm<PT: ProtocolTypes>:
    CodecP
    + Extractable<PT>
    + Debug
    + AsAny
    + 'static
where
    Self: 'static,
{
    // Required method
    fn boxed(&self) -> Box<dyn EvaluatedTerm<PT>>;

    // Provided methods
    fn type_id(&self) -> TypeId { ... }
    fn type_name(&self) -> &'static str { ... }
}

EvaluatedTerm: have both Codec and a way to extract knowledge out of a Message/OpaqueMessage or any type that might be used in a precomputation

Required Methods

fn boxed(&self) -> Box<dyn EvaluatedTerm<PT>>

Provided Methods

fn type_id(&self) -> TypeId

fn type_name(&self) -> &'static str