Trait puffin::protocol::ProtocolBehavior

pub trait ProtocolBehavior: 'static {
    type Claim: Claim;
    type SecurityViolationPolicy: SecurityViolationPolicy<Self::Claim>;
    type ProtocolMessage: ProtocolMessage<Self::OpaqueProtocolMessage>;
    type OpaqueProtocolMessage: OpaqueProtocolMessage;
    type Matcher: Matcher + for<'a> TryFrom<&'a MessageResult<Self::ProtocolMessage, Self::OpaqueProtocolMessage>>;

    fn signature() -> &'static Signature;
    fn registry() -> &'static PutRegistry<Self>
    where
        Self: Sized;
    fn create_corpus() -> Vec<(Trace<Self::Matcher>, &'static str)>;
}

Defines the protocol which is being tested. The fuzzer is generally abstract over the used protocol. We assume that protocols have opaque messages, structured messages, and a way to deframe an arbitrary stream of bytes into messages.

Also the library allows the definition of a type for claims and a (security policy)SecurityViolationPolicy over sequences of them. Finally, there is a matcher which allows traces to include queries for knowledge.

Required Associated Types

type Claim: Claim

type SecurityViolationPolicy: SecurityViolationPolicy<Self::Claim>

type ProtocolMessage: ProtocolMessage<Self::OpaqueProtocolMessage>

type OpaqueProtocolMessage: OpaqueProtocolMessage

type Matcher: Matcher + for<'a> TryFrom<&'a MessageResult<Self::ProtocolMessage, Self::OpaqueProtocolMessage>>

Required Methods

fn signature() -> &'static Signature

Get the signature which is used in the protocol

fn registry() -> &'static PutRegistry<Self> where
Self: Sized,

Gets the registry for concrete programs-under-test.

fn create_corpus() -> Vec<(Trace<Self::Matcher>, &'static str)>

Creates a sane initial seed corpus.