puffin::puffin_manager::file_accessor

Trait PuffinFileAccessor

pub trait PuffinFileAccessor:
    Send
    + Sync
    + 'static {
    type Reader: SizeAwareRangeReader + Sync;
    type Writer: AsyncWrite + Unpin + Send;
    type FileHandle: ToString + Clone + Send + Sync;

    // Required methods
    fn reader<'life0, 'life1, 'async_trait>(
        &'life0 self,
        handle: &'life1 Self::FileHandle,
    ) -> Pin<Box<dyn Future<Output = Result<Self::Reader>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn writer<'life0, 'life1, 'async_trait>(
        &'life0 self,
        handle: &'life1 Self::FileHandle,
    ) -> Pin<Box<dyn Future<Output = Result<Self::Writer>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

PuffinFileAccessor is for opening readers and writers for puffin files.

Required Associated Types

type Reader: SizeAwareRangeReader + Sync

type Writer: AsyncWrite + Unpin + Send

type FileHandle: ToString + Clone + Send + Sync

Required Methods

fn reader<'life0, 'life1, 'async_trait>( &'life0 self, handle: &'life1 Self::FileHandle, ) -> Pin<Box<dyn Future<Output = Result<Self::Reader>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Opens a reader for the given puffin file handle.

fn writer<'life0, 'life1, 'async_trait>( &'life0 self, handle: &'life1 Self::FileHandle, ) -> Pin<Box<dyn Future<Output = Result<Self::Writer>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Creates a writer for the given puffin file handle.