puffin::puffin_manager
Module stager

Modules

bounded_stager ๐

Structs

BoundedStager

BoundedStager is a Stager that uses moka to manage staging area.

FsBlobGuard

FsBlobGuard is a BlobGuard for accessing the blob and automatically deleting the file on drop.

FsDirGuard

FsDirGuard is a DirGuard for accessing the directory and automatically deleting the directory on drop.

Traits

DirWriterProvider

DirWriterProvider provides a way to write files into a directory.

Stager

Stager manages the staging area for the puffin files.

Type Aliases

BoxWriter

DirWriterProviderRef

WriteResult

Result containing the number of bytes written (u64).

Trait Aliases

InitBlobFn

Function that initializes a blob.

InitDirFn

Function that initializes a directory.