puffin::puffin_manager::stager

Module bounded_stager

Structs

BoundedStager

BoundedStager is a Stager that uses moka to manage staging area.

FsBlobGuard

FsBlobGuard is a BlobGuard for accessing the blob and automatically deleting the file on drop.

FsDirGuard

FsDirGuard is a DirGuard for accessing the directory and automatically deleting the directory on drop.

MokaDirWriterProvider 🔒

MokaDirWriterProvider implements DirWriterProvider for initializing a directory.

Enums

CacheValue 🔒

DeleteTask 🔒

Constants

DELETED_EXTENSION 🔒

DELETE_QUEUE_SIZE 🔒

RECYCLE_BIN_TTL 🔒

TMP_EXTENSION 🔒