Trait puffin::puffin_manager::stager::Stager

pub trait Stager: Send + Sync {
    type Blob: BlobGuard + Sync;
    type Dir: DirGuard;

    // Required methods
    fn get_blob<'a, 'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        puffin_file_name: &'life1 str,
        key: &'life2 str,
        init_factory: Box<dyn InitBlobFn + Send + Sync + 'a>
    ) -> Pin<Box<dyn Future<Output = Result<Self::Blob>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_dir<'a, 'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        puffin_file_name: &'life1 str,
        key: &'life2 str,
        init_fn: Box<dyn InitDirFn + Send + Sync + 'a>
    ) -> Pin<Box<dyn Future<Output = Result<Self::Dir>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn put_dir<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        puffin_file_name: &'life1 str,
        key: &'life2 str,
        dir_path: PathBuf,
        dir_size: u64
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
}

Stager manages the staging area for the puffin files.

Required Associated Types

type Blob: BlobGuard + Sync

type Dir: DirGuard

Required Methods

fn get_blob<'a, 'life0, 'life1, 'life2, 'async_trait>( &'life0 self, puffin_file_name: &'life1 str, key: &'life2 str, init_factory: Box<dyn InitBlobFn + Send + Sync + 'a> ) -> Pin<Box<dyn Future<Output = Result<Self::Blob>> + Send + 'async_trait>>
where Self: 'async_trait, 'a: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Retrieves a blob, initializing it if necessary using the provided init_fn.

The returned BlobGuard is used to access the blob reader. The caller is responsible for holding the BlobGuard until they are done with the blob.

fn get_dir<'a, 'life0, 'life1, 'life2, 'async_trait>( &'life0 self, puffin_file_name: &'life1 str, key: &'life2 str, init_fn: Box<dyn InitDirFn + Send + Sync + 'a> ) -> Pin<Box<dyn Future<Output = Result<Self::Dir>> + Send + 'async_trait>>
where Self: 'async_trait, 'a: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Retrieves a directory, initializing it if necessary using the provided init_fn.

The returned DirGuard is used to access the directory in the filesystem. The caller is responsible for holding the DirGuard until they are done with the directory.

fn put_dir<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, puffin_file_name: &'life1 str, key: &'life2 str, dir_path: PathBuf, dir_size: u64 ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Stores a directory in the staging area.