Trait puffin::puffin_manager::PuffinWriter

pub trait PuffinWriter {
    // Required methods
    fn put_blob<'life0, 'life1, 'async_trait, R>(
        &'life0 mut self,
        key: &'life1 str,
        raw_data: R,
        options: PutOptions
    ) -> Pin<Box<dyn Future<Output = Result<u64>> + Send + 'async_trait>>
       where R: AsyncRead + Send + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn put_dir<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        key: &'life1 str,
        dir: PathBuf,
        options: PutOptions
    ) -> Pin<Box<dyn Future<Output = Result<u64>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn set_footer_lz4_compressed(&mut self, lz4_compressed: bool);
    fn finish<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = Result<u64>> + Send + 'async_trait>>
       where Self: 'async_trait;
}

The PuffinWriter trait provides methods for writing blobs and directories to a Puffin file.

Required Methods

fn put_blob<'life0, 'life1, 'async_trait, R>( &'life0 mut self, key: &'life1 str, raw_data: R, options: PutOptions ) -> Pin<Box<dyn Future<Output = Result<u64>> + Send + 'async_trait>>
where R: AsyncRead + Send + 'async_trait, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Writes a blob associated with the specified key to the Puffin file. Returns the number of bytes written.

fn put_dir<'life0, 'life1, 'async_trait>( &'life0 mut self, key: &'life1 str, dir: PathBuf, options: PutOptions ) -> Pin<Box<dyn Future<Output = Result<u64>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Writes a directory associated with the specified key to the Puffin file. Returns the number of bytes written.

The specified dir should be accessible from the filesystem.

fn set_footer_lz4_compressed(&mut self, lz4_compressed: bool)

Sets whether the footer should be LZ4 compressed.

fn finish<'async_trait>( self ) -> Pin<Box<dyn Future<Output = Result<u64>> + Send + 'async_trait>>
where Self: 'async_trait,

Finalizes the Puffin file after writing.

Object Safety

This trait is not object safe.