Struct puffin::stream::MemoryStream

pub struct MemoryStream<D: ProtocolMessageDeframer> {
    inbound: Channel,
    outbound: Channel,
    deframer: D,
}

A MemoryStream has two Channels. The Stream also implements the Write and Read trait.

When writing to a MemoryStream its outbound channel gets filled.

When reading from a MemoryStream data is taken from the inbound channel.

This makes it possible for an crate::agent::Agent to treat a MemoryStream like a TLS socket! By writing to this socket you are sending data out. By reading from it you receive data.

Note: There need to be two separate buffer! Else for example a TLS socket would read and write into the same buffer

Fields
inbound: Channeloutbound: Channeldeframer: D
Implementations

impl<D: ProtocolMessageDeframer> MemoryStream<D>

pub fn new(deframer: D) -> Self