Struct puffin::trace::KnowledgeStore

pub struct KnowledgeStore<PB: ProtocolBehavior> {
    raw_knowledge: Vec<RawKnowledge<PB::Matcher>>,
}

Fields
raw_knowledge: Vec<RawKnowledge<PB::Matcher>>
Implementations

impl<PB: ProtocolBehavior> KnowledgeStore<PB>

pub fn new() -> Self

pub fn add_raw_knowledge<T: ExtractKnowledge<PB::Matcher> + 'static>( &mut self, data: T, source: Source, ) -> Result<usize, Error>

pub fn number_matching_message_with_source( &self, source: Source, type_id: TypeId, tls_message_type: &Option<PB::Matcher>, ) -> usize

pub fn number_matching_message( &self, type_id: TypeId, tls_message_type: &Option<PB::Matcher>, ) -> usize

Count the number of sub-messages of type type_id in the output message.

pub fn find_variable( &self, query_type_shape: TypeShape, query: &Query<PB::Matcher>, ) -> Option<&dyn VariableData>

Returns the variable which matches best -> highest specificity If we want a variable with lower specificity, then we can just query less specific