Struct puffin::trace::Trace

pub struct Trace<M: Matcher> {
    pub descriptors: Vec<AgentDescriptor>,
    pub steps: Vec<Step<M>>,
    pub prior_traces: Vec<Trace<M>>,
}

Fields
descriptors: Vec<AgentDescriptor>steps: Vec<Step<M>>prior_traces: Vec<Trace<M>>
Implementations

impl<M: Matcher> Trace<M>

pub fn dot_graph(&self, tree_mode: bool) -> String

pub fn dot_subgraphs(&self, tree_mode: bool) -> Vec<String>

impl<M: Matcher> Trace<M>

pub fn count_functions_by_name(&self, find_name: &'static str) -> usize

pub fn count_functions(&self) -> usize

pub fn write_plots(&self, i: u16)

impl<M: Matcher> Trace<M>

A Trace consists of several Steps. Each has either a OutputAction or an InputAction. Each Steps references an Agent by name. Furthermore, a trace also has a list of AgentDescriptors which act like a blueprint to spawn Agents with a corresponding server or client role and a specific TLs version. Essentially they are an Agent without a stream.

fn spawn_agents<PB: ProtocolBehavior>(
&self,
ctx: &mut TraceContext<PB>
) -> Result<(), Error>

pub fn execute<PB>(&self, ctx: &mut TraceContext<PB>) -> Result<(), Error>where
PB: ProtocolBehavior<Matcher = M>,

pub fn execute_deterministic<PB>(
&self,
put_registry: &PutRegistry<PB>,
default_put_options: PutOptions
) -> Result<TraceContext<PB>, Error>where
PB: ProtocolBehavior<Matcher = M>,

pub fn execute_with_non_default_puts<PB>(
&self,
put_registry: &PutRegistry<PB>,
descriptors: &[(AgentName, PutDescriptor)]
) -> Result<TraceContext<PB>, Error>where
PB: ProtocolBehavior<Matcher = M>,

pub fn serialize_postcard(&self) -> Result<Vec<u8>, Error>

pub fn deserialize_postcard(slice: &[u8]) -> Result<Trace<M>, Error>