Struct puffin::trace::TraceContext

pub struct TraceContext<PB: ProtocolBehavior> {
    knowledge: Vec<Knowledge<PB::Matcher>>,
    agents: Vec<Agent<PB>>,
    claims: GlobalClaimList<<PB as ProtocolBehavior>::Claim>,
    put_registry: PutRegistry<PB>,
    deterministic_put: bool,
    default_put_options: PutOptions,
    non_default_put_descriptors: HashMap<AgentName, PutDescriptor>,
    phantom: PhantomData<PB>,
}

The TraceContext contains a list of VariableData, which is known as the knowledge of the attacker. VariableData can contain data of various types like for example client and server extensions, cipher suits or session ID It also holds the concrete references to the Agents and the underlying streams, which contain the messages which have need exchanged and are not yet processed by an output step.

Fields
knowledge: Vec<Knowledge<PB::Matcher>>

The knowledge of the attacker

agents: Vec<Agent<PB>>claims: GlobalClaimList<<PB as ProtocolBehavior>::Claim>put_registry: PutRegistry<PB>deterministic_put: booldefault_put_options: PutOptionsnon_default_put_descriptors: HashMap<AgentName, PutDescriptor>phantom: PhantomData<PB>
Implementations

impl<PB: ProtocolBehavior> TraceContext<PB>

pub fn new(
put_registry: &PutRegistry<PB>,
default_put_options: PutOptions
) -> Self

pub fn put_registry(&self) -> &PutRegistry<PB>

pub fn claims(&self) -> &GlobalClaimList<PB::Claim>

pub fn verify_security_violations(&self) -> Result<(), Error>

pub fn add_knowledge(&mut self, knowledge: Knowledge<PB::Matcher>)

pub fn number_matching_message(
&self,
agent: AgentName,
type_id: TypeId,
tls_message_type: &Option<PB::Matcher>
) -> usize

Count the number of sub-messages of type [type_id] in the output message [in_step_id].

pub fn find_claim(
&self,
agent_name: AgentName,
query_type_shape: TypeShape
) -> Option<Box<dyn Any>>

pub fn find_variable(
&self,
query_type_shape: TypeShape,
query: &Query<PB::Matcher>
) -> Option<&dyn VariableData>

Returns the variable which matches best -> highest specificity If we want a variable with lower specificity, then we can just query less specific

pub fn add_to_inbound(
&mut self,
agent_name: AgentName,
message: &PB::OpaqueProtocolMessage
) -> Result<(), Error>

Adds data to the inbound Channel of the Agent referenced by the parameter “agent”.

pub fn next_state(&mut self, agent_name: AgentName) -> Result<(), Error>

pub fn take_message_from_outbound(
&mut self,
agent_name: AgentName
) -> Result<Option<MessageResult<PB::ProtocolMessage, PB::OpaqueProtocolMessage>>, Error>

Takes data from the outbound Channel of the Agent referenced by the parameter “agent”. See [MemoryStream::take_message_from_outbound]

fn add_agent(&mut self, agent: Agent<PB>) -> AgentName

pub fn new_agent(
&mut self,
descriptor: &AgentDescriptor
) -> Result<AgentName, Error>

pub fn find_agent_mut(
&mut self,
name: AgentName
) -> Result<&mut Agent<PB>, Error>

pub fn find_agent(&self, name: AgentName) -> Result<&Agent<PB>, Error>

pub fn put_descriptor(&self, agent_descriptor: &AgentDescriptor) -> PutDescriptor

Gets the PUT descriptor which should be used for all agents

fn default_put_descriptor(&self) -> PutDescriptor

pub fn set_non_default_put(
&mut self,
agent_name: AgentName,
put_descriptor: PutDescriptor
)

Makes agents use the non-default PUT

pub fn set_non_default_puts(
&mut self,
descriptors: &[(AgentName, PutDescriptor)]
)

pub fn reset_agents(&mut self) -> Result<(), Error>

pub fn agents_successful(&self) -> bool

pub fn set_deterministic(&mut self, deterministic: bool)