Crate puffin

Usage:

fn main() {
    puffin::set_scopes_on(true); // you may want to control this with a flag

    // game loop
    loop {
        puffin::GlobalProfiler::lock().new_frame();

        {
            puffin::profile_scope!("slow_code");
            slow_code();
        }

    }
}

Macros

current_file_name

Returns a shortened path to the current file.

current_function_name

Returns the name of the calling function without a long module path prefix.

profile_function

Automatically name the profiling scope based on function name.

profile_scope

Profile the current scope with the given name (unique in the parent scope).

Structs

FrameData

One frame worth of profile data, collected from many sources.

FrameMeta

Meta-information about a frame.

FrameSinkId

Identifies a specific FrameSink when added to GlobalProfiler.

FrameView

A view of recent and slowest frames, used by GUIs.

GlobalFrameView

Automatically connects to crate::GlobalProfiler.

GlobalProfiler

Singleton. Collects profiling data from multiple threads and passes them on to different FrameSinks.

MergeScope

A scope that has been merged from many different sources

ProfilerScope

Created by the puffin::profile*!(...) macros.

Reader

Parses a Stream of profiler data.

Record

Used when parsing a Stream.

Scope

Used when parsing a Stream.

Stream

Stream of profiling events from one thread.

StreamInfo

A Stream plus some info about it.

StreamInfoRef

A reference to the contents of a StreamInfo.

ThreadInfo

Used to identify one source of profiling data.

ThreadProfiler

Collects profiling data for one thread

UnpackedFrameData

One frame worth of profile data, collected from many sources.

Enums

Error

Errors that can happen when parsing a Stream of profile data.

Functions

are_scopes_on

Are the profiler scope macros turned on? This is false by default.

global_reporter

Report a stream of profile data from a thread to the GlobalProfiler singleton.

merge_scopes_for_thread

For the given thread, merge all scopes with the same id path.

now_ns

Returns a high-precision, monotonically increasing nanosecond count since unix epoch.

select_slowest

Select the slowest frames, up to a certain count.

set_scopes_on

Turn on/off the profiler macros (profile_function, profile_scope etc). When off, these calls take only 1-2 ns to call (100x faster). This is false by default.

Type Definitions

FrameIndex

FrameSink

Add these to GlobalProfiler with GlobalProfiler::add_sink().

NanoSecond

All times are expressed as integer nanoseconds since some event.

Result