Crate puffin

Re-exports

pub use libafl;

pub use libafl_bolts;

Modules

agent

Agents represent communication participants like Alice, Bob or Eve.

algebra

The term module defines typedDYTerms of the form fn_add(x: u8, fn_square(y: u16)) → u16.

claims

cli

codec

error

execution

experiment

fuzzer

The fuzzer module setups the fuzzing loop. It also is responsible for gathering feedback from runs and restarting processes if they crash.

graphviz

This module adds plotting capabilities toDYTermsand Traces. The output of the functions in this module can be passed to the command line utility dot which is part of graphviz.

harness

log

protocol

put

put_registry

stream

The communication streams between Agents.

test_utils

trace

This module define the execution Traces.

trace_helper

Macros

atom_extract_knowledge

define_signature

dummy_codec

dummy_extract_knowledge

dummy_extract_knowledge_codec

expand_mutations

input_action

This macro defines the precomputation syntax to add precomputations to an input action step

supports

term

term_arg

test_puts