Module puffin::agent

Agents represent communication participants like Alice, Bob or Eve.

Note that attackers are usually not represented by these Agents but instead through a recipe term (see crate::trace::InputAction).

Each Agent has an inbound and an outbound channel (see crate::stream)

Structs

Agent

An Agent holds a non-cloneable reference to a Stream.

AgentDescriptor

AgentDescriptors act like a blueprint to spawn Agents with a corresponding server or client role and a specific TLs version. Essentially they are an Agent without a stream.

AgentName

Copyable reference to an Agent. It identifies exactly one agent.

Enums

AgentType

TLSVersion