Enum puffin::algebra::term::Term

pub enum Term<M: Matcher> {
    Variable(Variable<M>),
    Application(Function, Vec<Term<M>>),
}

A first-order term: either a Variable or an application of an Function.

Variants

Variable(Variable<M>)

A concrete but unspecified Term (e.g. x, y). See Variable for more information.

Application(Function, Vec<Term<M>>)

An Function applied to zero or more Terms (e.g. (f(x, y), g()).

A Term that is an application of an Function with arity 0 applied to 0 Terms can be considered a constant.

Implementations

impl<M: Matcher> Term<M>

pub fn resistant_id(&self) -> u32

pub fn size(&self) -> usize

pub fn is_leaf(&self) -> bool

pub fn get_type_shape(&self) -> &TypeShape

pub fn name(&self) -> &str

pub fn mutate(&mut self, other: Term<M>)

fn display_at_depth(&self, depth: usize) -> String

pub fn evaluate<PB>( &self, context: &TraceContext<PB>, ) -> Result<Box<dyn Any>, Error>
where PB: ProtocolBehavior<Matcher = M>,

impl<M: Matcher> Term<M>

fn unique_id(&self, tree_mode: bool, cluster_id: usize) -> String

fn node_attributes( displayable: impl Display, color: &str, shape: &str, ) -> String

fn collect_statements( term: &Term<M>, tree_mode: bool, cluster_id: usize, statements: &mut Vec<String>, )

pub fn dot_subgraph( &self, tree_mode: bool, cluster_id: usize, label: &str, ) -> String

If tree_mode is true then each subgraph is self-contained and does not reference other clusters or nodes outside of this subgraph. Therefore, only trees are generated. If it is false, then graphs are rendered.

impl<M: Matcher> Term<M>

pub fn count_functions_by_name(&self, find_name: &'static str) -> usize