Trait puffin::codec::Codec

pub trait Codec: Debug + Sized {
    // Required methods
    fn encode(&self, bytes: &mut Vec<u8>);
    fn read(_: &mut Reader<'_>) -> Option<Self>;

    // Provided methods
    fn get_encoding(&self) -> Vec<u8>  { ... }
    fn read_bytes(bytes: &[u8]) -> Option<Self> { ... }
}

Things we can encode and read from a Reader.

Required Methods

fn encode(&self, bytes: &mut Vec<u8>)

Encode yourself by appending onto bytes.

fn read(_: &mut Reader<'_>) -> Option<Self>

Decode yourself by fiddling with the Reader. Return Some if it worked, None if not.

Provided Methods

fn get_encoding(&self) -> Vec<u8>

Convenience function to get the results of encode().

fn read_bytes(bytes: &[u8]) -> Option<Self>

Read one of these from the front of bytes and return it.

Object Safety

This trait is not object safe.