Struct puffin::partial_reader::PartialReader

pub struct PartialReader<R> {
    offset: u64,
    size: u64,
    source: R,
    position_in_portion: Option<u64>,
}

PartialReader to perform synchronous or asynchronous reads on a portion of a resource.

Fields
offset: u64

offset of the portion in the resource

size: u64

size of the portion in the resource

source: R

Resource for the portion. The offset and size fields are used to determine the slice of source to read.

position_in_portion: Option<u64>

The current position within the portion.

A None value indicates that no read operations have been performed yet on this portion. Before a read operation can be performed, the resource must be positioned at the correct offset in the portion. After the first read operation, this field will be set to Some(_), representing the current read position in the portion.

Implementations

impl<R> PartialReader<R>

pub(crate) fn project<'pin>( self: Pin<&'pin mut Self> ) -> __PartialReaderProjection<'pin, R>

pub(crate) fn project_ref<'pin>( self: Pin<&'pin Self> ) -> __PartialReaderProjectionRef<'pin, R>

impl<R> PartialReader<R>

pub fn new(source: R, offset: u64, size: u64) -> Self

Creates a new PartialReader for the given resource.

pub fn position(&self) -> u64

Returns the current position in the portion.

pub fn size(&self) -> u64

Returns the size of the portion in portion.

pub fn is_empty(&self) -> bool

Returns whether the portion is empty.

pub fn is_eof(&self) -> bool

Returns whether the current position is at the end of the portion.