puffin

Module puffin_manager

Modules

cache

file_accessor

fs_puffin_manager

stager

Structs

GuardWithMetadata

GuardWithMetadata provides access to the blob or directory data and its metadata.

PutOptions

Options available for put_blob and put_dir methods.

Traits

BlobGuard

BlobGuard is provided by the PuffinReader to access the blob data. Users should hold the BlobGuard until they are done with the blob data.

DirGuard

DirGuard is provided by the PuffinReader to access the directory in the filesystem. Users should hold the DirGuard until they are done with the directory.

PuffinManager

The PuffinManager trait provides a unified interface for creating PuffinReader and PuffinWriter.

PuffinReader

The PuffinReader trait provides methods for reading blobs and directories from a Puffin file.

PuffinWriter

The PuffinWriter trait provides methods for writing blobs and directories to a Puffin file.