Trait puffin::puffin_manager::PuffinReader

pub trait PuffinReader {
    type Blob: BlobGuard;
    type Dir: DirGuard;

    // Required methods
    fn blob<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Self::Blob>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn dir<'life0, 'life1, 'async_trait>(
        &'life0 self,
        key: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Self::Dir>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

The PuffinReader trait provides methods for reading blobs and directories from a Puffin file.

Required Associated Types

type Blob: BlobGuard

type Dir: DirGuard

Required Methods

fn blob<'life0, 'life1, 'async_trait>( &'life0 self, key: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<Self::Blob>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Reads a blob from the Puffin file.

The returned BlobGuard is used to access the blob data. Users should hold the BlobGuard until they are done with the blob data.

fn dir<'life0, 'life1, 'async_trait>( &'life0 self, key: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<Self::Dir>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Reads a directory from the Puffin file.

The returned DirGuard is used to access the directory in the filesystem. The caller is responsible for holding the DirGuard until they are done with the directory.