Trait puffin::put::Put

pub trait Put<PB: ProtocolBehavior>: Stream<PB::ProtocolMessage, PB::OpaqueProtocolMessage> + 'static {
    fn progress(&mut self, agent_name: &AgentName) -> Result<(), Error>;
    fn reset(&mut self, agent_name: AgentName) -> Result<(), Error>;
    fn descriptor(&self) -> &AgentDescriptor;
    fn rename_agent(&mut self, agent_name: AgentName) -> Result<(), Error>;
    fn describe_state(&self) -> &str;
    fn is_state_successful(&self) -> bool;
    fn determinism_reseed(&mut self) -> Result<(), Error>;
    fn shutdown(&mut self) -> String;
    fn version() -> String
    where
        Self: Sized;

    fn is_reusable_with(&self, other: &AgentDescriptor) -> bool { ... }
}

Defines the interface which all programs-under-test need to implement. They need a way to progress, reset and describe their state.

Required Methods

fn progress(&mut self, agent_name: &AgentName) -> Result<(), Error>

Process incoming buffer, internal progress, can fill in output buffer

fn reset(&mut self, agent_name: AgentName) -> Result<(), Error>

In-place reset of the state

fn descriptor(&self) -> &AgentDescriptor

fn rename_agent(&mut self, agent_name: AgentName) -> Result<(), Error>

Propagate agent changes to the PUT

fn describe_state(&self) -> &str

Returns a textual representation of the state in which self is

fn is_state_successful(&self) -> bool

Checks whether the Put is in a good state

fn determinism_reseed(&mut self) -> Result<(), Error>

Make the PUT used by self determimistic in the future by making its PRNG “deterministic” Now subsumed by Factory-level functions to reseed globally: determinism_reseed

fn shutdown(&mut self) -> String

Shutdown the PUT by consuming it and returning a string which summarizes the execution.

fn version() -> Stringwhere
Self: Sized,

Returns a textual representation of the version of the PUT used by self

Provided Methods

fn is_reusable_with(&self, other: &AgentDescriptor) -> bool

checks whether a agent is reusable with the descriptor