Trait puffin::put::Put

pub trait Put<PB: ProtocolBehavior>: Stream<PB::Matcher, PB::ProtocolMessage, PB::OpaqueProtocolMessage, PB::OpaqueProtocolMessageFlight> + 'static {
    // Required methods
    fn progress(&mut self, agent_name: &AgentName) -> Result<(), Error>;
    fn reset(&mut self, agent_name: AgentName) -> Result<(), Error>;
    fn descriptor(&self) -> &AgentDescriptor;
    fn rename_agent(&mut self, agent_name: AgentName) -> Result<(), Error>;
    fn describe_state(&self) -> &str;
    fn is_state_successful(&self) -> bool;
    fn determinism_reseed(&mut self) -> Result<(), Error>;
    fn shutdown(&mut self) -> String;
    fn version() -> String
       where Self: Sized;

    // Provided method
    fn is_reusable_with(&self, other: &AgentDescriptor) -> bool { ... }
}

Generic trait used to define the interface with a concrete library implementing the protocol.

Required Methods

fn progress(&mut self, agent_name: &AgentName) -> Result<(), Error>

Process incoming buffer, internal progress, can fill in the output buffer

fn reset(&mut self, agent_name: AgentName) -> Result<(), Error>

In-place reset of the state

fn descriptor(&self) -> &AgentDescriptor

fn rename_agent(&mut self, agent_name: AgentName) -> Result<(), Error>

Propagate agent changes to the PUT

fn describe_state(&self) -> &str

Returns a textual representation of the state in which self is

fn is_state_successful(&self) -> bool

Checks whether the Put is in a good state

fn determinism_reseed(&mut self) -> Result<(), Error>

Make the PUT used by self deterministic in the future by making its PRNG “deterministic” Now subsumed by Factory-level functions to reseed globally: determinism_reseed

fn shutdown(&mut self) -> String

Shut down the PUT by consuming it and returning a string that summarizes the execution.

fn version() -> String
where Self: Sized,

Returns a textual representation of the version of the PUT used by self

Provided Methods

fn is_reusable_with(&self, other: &AgentDescriptor) -> bool

checks whether a agent is reusable with the descriptor