Struct puffin::trace::Trace

pub struct Trace<PT: ProtocolTypes> {
    pub descriptors: Vec<AgentDescriptor>,
    pub steps: Vec<Step<PT>>,
    pub prior_traces: Vec<Trace<PT>>,
}

Fields
descriptors: Vec<AgentDescriptor>steps: Vec<Step<PT>>prior_traces: Vec<Trace<PT>>
Implementations

impl<PT: ProtocolTypes> Trace<PT>

pub fn dot_graph(&self, tree_mode: bool) -> String

pub fn dot_subgraphs(&self, tree_mode: bool) -> Vec<String>

impl<PT: ProtocolTypes> Trace<PT>

pub fn count_functions_by_name(&self, find_name: &'static str) -> usize

pub fn count_functions(&self) -> usize

pub fn write_plots(&self, i: u16)

impl<PT: ProtocolTypes> Trace<PT>

A Trace consists of several Steps. Each has either a OutputAction or an InputAction. Each Steps references an Agent by name. Furthermore, a trace also has a list of AgentDescriptors which act like a blueprint to spawn Agents with a corresponding server or client role and a specific TLs version. Essentially they are an Agent without a stream.

pub fn spawn_agents<PB: ProtocolBehavior>( &self, ctx: &mut TraceContext<PB>, ) -> Result<(), Error>

pub fn execute_until_step<PB>( &self, ctx: &mut TraceContext<PB>, nb_steps: usize, ) -> Result<(), Error>
where PB: ProtocolBehavior<ProtocolTypes = PT>,

pub fn execute<PB>(&self, ctx: &mut TraceContext<PB>) -> Result<(), Error>
where PB: ProtocolBehavior<ProtocolTypes = PT>,

pub fn serialize_postcard(&self) -> Result<Vec<u8>, Error>

pub fn deserialize_postcard(slice: &[u8]) -> Result<Self, Error>

pub fn all_payloads(&self) -> Vec<&Payloads>

pub fn all_payloads_mut(&mut self) -> Vec<&mut Payloads>

pub fn is_symbolic(&self) -> bool