Struct puffin::trace::TraceContext

pub struct TraceContext<PB: ProtocolBehavior> {
    pub knowledge_store: KnowledgeStore<PB::ProtocolTypes>,
    agents: Vec<Agent<PB>>,
    claims: GlobalClaimList<PB::ProtocolTypes, PB::Claim>,
    spawner: Spawner<PB>,
    phantom: PhantomData<PB>,
}

The TraceContext represents the state of an execution.

The TraceContext contains a list of VariableData, which is known as the knowledge of the attacker. VariableData can contain data of various types like for example client and server extensions, cipher suits or session ID It also holds the concrete references to the Agents and the underlying streams, which contain the messages which have need exchanged and are not yet processed by an output step.

Fields
knowledge_store: KnowledgeStore<PB::ProtocolTypes>

The knowledge of the attacker

agents: Vec<Agent<PB>>claims: GlobalClaimList<PB::ProtocolTypes, PB::Claim>spawner: Spawner<PB>phantom: PhantomData<PB>
Implementations

impl<PB: ProtocolBehavior> TraceContext<PB>

pub fn new(spawner: Spawner<PB>) -> Self

pub fn verify_security_violations(&self) -> Result<(), Error>

pub fn number_matching_message_with_source( &self, source: Source, type_id: TypeId, tls_message_type: &Option<<PB::ProtocolTypes as ProtocolTypes>::Matcher>, ) -> usize

Count the number of sub-messages of type type_id with the correct source

pub fn number_matching_message( &self, type_id: TypeId, tls_message_type: &Option<<PB::ProtocolTypes as ProtocolTypes>::Matcher>, ) -> usize

Count the number of sub-messages of type type_id in the output message.

pub fn find_claim( &self, agent_name: AgentName, query_type_shape: TypeShape<PB::ProtocolTypes>, ) -> Option<Box<dyn EvaluatedTerm<PB::ProtocolTypes>>>

pub fn find_variable( &self, query_type_shape: TypeShape<PB::ProtocolTypes>, query: &Query<<PB::ProtocolTypes as ProtocolTypes>::Matcher>, ) -> Option<&dyn VariableData<PB::ProtocolTypes>>

Returns the variable which matches best -> highest specificity If we want a variable with lower specificity, then we can just query less specific

pub fn spawn(&mut self, descriptor: &AgentDescriptor) -> Result<(), Error>

pub fn find_agent_mut( &mut self, name: AgentName, ) -> Result<&mut Agent<PB>, Error>

pub fn find_agent(&self, name: AgentName) -> Result<&Agent<PB>, Error>

pub fn agents_successful(&self) -> bool